#5 리스트에서 롱클릭시 확인없이 삭제 help Wndted
해결
RecyclerViewAdapter의 OnItemLongClickListener의 메서드 itemLongClick 의 return 값 true 라면 삭제 false라면 삭제
취소로 지정하여
MainActivity의 AlertDialog의 버튼 신호에 따라 삭제여부 확인

+++ app/src/main/java/com/example/user/notepad/DBHelper.java
@@ -109,15 +109,21 @@ public class DBHelper extends SQLiteOpenHelper {
     }
 
 
-    void deleteDataTableIndex(String data) {
-        System.out.println(data);
-        db = getWritableDatabase();
-        String sql = "DELETE FROM " + TABLE_NAME +
-                " WHERE Data = '" + data + "'; ";
-        db.execSQL(sql);
-        db.close();
+    boolean deleteDataTableIndex(String data) {
+        if(!data.equals("")) {
+            System.out.println(data);
+            db = getWritableDatabase();
+            String sql = "DELETE FROM " + TABLE_NAME +
+                    " WHERE Data = '" + data + "'; ";
+            db.execSQL(sql);
+            db.close();
 
-        Log.i(CLASS_NAME, "delete Sucess");
+            Log.i(CLASS_NAME, "delete Sucess");
+
+            return true;
+        }
+        else
+            return false;
     }
 
     void selectDataTableIndex() {
